refactor: extract disk module from volumes
Move disk image operations to dedicated `disk/` module:
- disk.rs - Disk struct, DiskFormat enum
- ext4.rs - Ext4 filesystem creation
- qcow2.rs - QCOW2 copy-on-write operations
- constants.rs - Disk size constants

The `volumes/` module now focuses on VMM volume configuration
(BlockDeviceManager), while `disk/` handles disk image operations.

This separation improves code organization and makes the module
responsibilities clearer.
