+++ boxlite/src/disk/image.rs
