+++ boxlite/src/images/store.rs
@@ -194,10 +194,10 @@ impl ImageStore {
     ///
     /// Returns a persistent Disk if the cached disk image exists, None otherwise.
     /// The returned Disk is persistent (won't be deleted on drop).
-    pub async fn disk_image(&self, image_digest: &str) -> Option<crate::volumes::Disk> {
+    pub async fn disk_image(&self, image_digest: &str) -> Option<crate::disk::Disk> {
         let inner = self.inner.read().await;
         if let Some((path, format)) = inner.storage.find_disk_image(image_digest) {
-            Some(crate::volumes::Disk::new(path, format, true))
+            Some(crate::disk::Disk::new(path, format, true))
         } else {
             None
         }
@@ -218,8 +218,8 @@ impl ImageStore {
     pub async fn install_disk_image(
         &self,
         image_digest: &str,
-        disk: crate::volumes::Disk,
-    ) -> BoxliteResult<crate::volumes::Disk> {
+        disk: crate::disk::Disk,
+    ) -> BoxliteResult<crate::disk::Disk> {
         let inner = self.inner.read().await;
         let disk_format = disk.format();
         let target_path = inner.storage.disk_image_path(image_digest, disk_format);
@@ -240,7 +240,7 @@ impl ImageStore {
             tracing::debug!("Disk image already installed: {}", target_path.display());
             // Leak the source disk to prevent cleanup (it may have been the same file)
             let _ = disk.leak();
-            return Ok(crate::volumes::Disk::new(target_path, disk_format, true));
+            return Ok(crate::disk::Disk::new(target_path, disk_format, true));
         }
 
         let source_path = disk.path().to_path_buf();
@@ -264,7 +264,7 @@ impl ImageStore {
             target_path.display()
         );
 
-        Ok(crate::volumes::Disk::new(target_path, disk_format, true))
+        Ok(crate::disk::Disk::new(target_path, disk_format, true))
     }
 
     // ========================================================================
